chore: split manage-account tests into 3 files
